pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "essaywriter-pdf-pipeline"
version = "0.1.0"
description = "Baseline PDF text extraction pipeline."
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
  "pypdf>=4.2.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
ocr-small = [
  "pypdfium2>=4.30.0",
  "pytesseract>=0.3.10",
]
ocr-medium = [
  "pypdfium2>=4.30.0",
  "easyocr>=1.7.0",
  "numpy>=1.24.0",
]
ocr-high = [
  "pypdfium2>=4.30.0",
  "paddleocr>=2.8.0",
  "paddlepaddle>=2.6.0",
  "numpy>=1.24.0",
]
ocr-scheduler = ["psutil>=5.9.0"]
llm-claude = ["anthropic>=0.39.0"]
llm-openai = ["openai>=1.54.0"]
llm-gemini = ["google-generativeai>=0.8.0"]
llm-all = ["anthropic>=0.39.0", "openai>=1.54.0", "google-generativeai>=0.8.0"]
outline = ["rapidfuzz>=3.10.0", "pypdf>=5.0.0", "jsonschema>=4.21.0"]

[project.scripts]
pdf-extract = "pdf_pipeline.cli:main"

[tool.setuptools.packages.find]
include = ["pdf_pipeline*", "llm*", "essay_writer*"]
exclude = ["tests*", "testpdfs*", "docs*"]

[tool.pytest.ini_options]
pythonpath = ["."]
